Create program2.cpp
On a certain day the British pound was equivalent to $1.487 U.S., the French franc was $0.172, the German deutsche mark was $0.584, and the Japanese yen was $0.00955. Write a program that allows the user to enter an amount in dollars, and then displays this value converted to these four other monetary units.