Create program6.cpp
Create a structure called Volume that uses three variables of type Distance (from the ENGLSTRC example) to model the volume of a room. 
Initialize a variable of type Volume to specific dimensions, then calculate the volume it represents, and print out the result.
To calculate the volume, convert each dimension from a Distance variable to a variable
of type float representing feet and fractions of a foot, and then multiply the resulting three numbers.